try huhuhu

+++ libs/baseclass/calibrate_bg.py
@@ -104,7 +104,7 @@ class CalibrateBG(Screen):
         self.ids['capture_bg'].disabled = not self.ids['capture_bg'].disabled
 
     def disable_clock(self):
-        self.insert_data('background', np.array(self.spec.intensities(False,True), dtype=np.float64).reshape(-1, 1))
+        self.insert_data('background', np.array(self.spec.intensities(False,True), dtype=np.float32).reshape(-1, 1))
         Clock.unschedule(self.update_graph)
     
     def on_leave(self, *args):
